[refactor] visits api url 변경
- visits api url 변경
  기존 api를 남겨두기 위해 임시로 사용했던 visitsMulti 이름 전부 삭제
