+++ src/routes/visits/controllers/getVisits.ts
@@ -1,4 +1,5 @@
 import { Context } from 'koa';
+import convertToArray from '../../../utils/convertToArray';
 import {
   getDailyInMonth,
   getMonthlyInYear,
@@ -6,31 +7,28 @@ import {
   fillMonthCountsWithZero,
 } from '../services/visitsService';
 
-interface IParams {
-  projectId: string;
-}
-
-interface IQuery {
-  type: string;
-  year: number;
-  month: number;
-}
-
 export default async (ctx: Context): Promise<void> => {
-  const { projectId }: IParams = ctx.params;
-  const { type, year, month }: IQuery = ctx.query;
+  const { projectId, type, year, month } = ctx.query;
   if (type === 'daily') {
-    const visitsByDate = await getDailyInMonth({ projectId, year, month });
-    const filledData = fillDateCountsWithZero(visitsByDate, year, month);
-    ctx.response.body = filledData;
-    return;
+    const projectIds = convertToArray(projectId);
+    const filledDatas = await Promise.all(
+      projectIds.map(async (targetProjectId) => {
+        const visitsByDate = await getDailyInMonth({ targetProjectId, year, month });
+        const filledData = fillDateCountsWithZero({ visitsByDate, targetProjectId, year, month });
+        return filledData;
+      }),
+    );
+    ctx.response.body = filledDatas;
   }
   if (type === 'monthly') {
-    const visitsByMonth = await getMonthlyInYear({ projectId, year });
-    const filledData = fillMonthCountsWithZero(visitsByMonth, year);
-    ctx.response.body = filledData;
-    return;
+    const projectIds = convertToArray(projectId);
+    const filledDatas = await Promise.all(
+      projectIds.map(async (targetProjectId) => {
+        const visitsByMonth = await getMonthlyInYear({ targetProjectId, year });
+        const filledData = fillMonthCountsWithZero({ visitsByMonth, targetProjectId, year });
+        return filledData;
+      }),
+    );
+    ctx.response.body = filledDatas;
   }
-
-  ctx.throw(400, 'internal server error');
 };
