+++ src/routes/visits/services/visitsService.ts
@@ -2,22 +2,21 @@ import { Types } from 'mongoose';
 import Visits, { IVisitsDocument } from '../../../models/Visits';
 
 interface IDailyInMonthParams {
-  projectId: string;
+  targetProjectId: string;
   year: number;
   month: number;
 }
 
 const getDailyInMonth = async (params: IDailyInMonthParams): Promise<IVisitsDocument[]> => {
-  const { projectId, year, month }: IDailyInMonthParams = params;
+  const { targetProjectId, year, month }: IDailyInMonthParams = params;
 
   const startDate = new Date(year, month - 1, 1);
   const endDate = new Date(year, month, 0);
-
   const visitsByDate = await Visits.aggregate([
     {
       $match: {
         $and: [
-          { projectId: Types.ObjectId(projectId) },
+          { projectId: Types.ObjectId(targetProjectId) },
           { date: { $gte: startDate, $lte: endDate } },
         ],
       },
@@ -25,6 +24,7 @@ const getDailyInMonth = async (params: IDailyInMonthParams): Promise<IVisitsDocu
     {
       $group: {
         _id: {
+          projectId: '$projectId',
           year: { $year: '$date' },
           month: { $month: '$date' },
           date: { $dayOfMonth: '$date' },
@@ -37,48 +37,15 @@ const getDailyInMonth = async (params: IDailyInMonthParams): Promise<IVisitsDocu
   return visitsByDate;
 };
 
-interface IMonthlyInYearParams {
-  projectId: string;
+interface FillDateCountParams {
+  visitsByDate: IVisitsDocument[];
+  targetProjectId: string;
   year: number;
+  month: number;
 }
 
-const getMonthlyInYear = async (params: IMonthlyInYearParams): Promise<IVisitsDocument[]> => {
-  const { projectId, year }: IMonthlyInYearParams = params;
-  const startDate = new Date(year, 0, 1);
-  const endDate = new Date(year, 12, 0);
-  const visitsByMonth = await Visits.aggregate([
-    {
-      $match: {
-        $and: [
-          { projectId: Types.ObjectId(projectId) },
-          { date: { $gte: startDate, $lte: endDate } },
-        ],
-      },
-    },
-    {
-      $group: {
-        _id: {
-          year: { $year: '$date' },
-          month: { $month: '$date' },
-          ip: '$ip',
-        },
-      },
-    },
-    {
-      $group: {
-        _id: {
-          year: '$_id.year',
-          month: '$_id.month',
-        },
-        count: { $sum: 1 },
-      },
-    },
-    { $sort: { '_id.month': 1 } },
-  ]);
-  return visitsByMonth;
-};
-
-const fillDateCountsWithZero = (visitsByDate: IVisitsDocument[], year: number, month: number) => {
+const fillDateCountsWithZero = (params: FillDateCountParams) => {
+  const { visitsByDate, targetProjectId: projectId, year, month } = params;
   const lastDate = new Date(year, month + 1, 0).getDate();
   let i = 0;
   const filledVisitsByMonth = Array(lastDate)
@@ -91,7 +58,7 @@ const fillDateCountsWithZero = (visitsByDate: IVisitsDocument[], year: number, m
         return targetData;
       }
       const zeroData = {
-        _id: { year, month, date: targetDate },
+        _id: { year, month, date: targetDate, projectId },
         count: 0,
       };
       return zeroData;
@@ -99,7 +66,57 @@ const fillDateCountsWithZero = (visitsByDate: IVisitsDocument[], year: number, m
   return filledVisitsByMonth;
 };
 
-const fillMonthCountsWithZero = (visitsByMonth: IVisitsDocument[], year: number) => {
+interface IMonthlyInYearParams {
+  targetProjectId: string;
+  year: number;
+}
+
+const getMonthlyInYear = async (params: IMonthlyInYearParams): Promise<IVisitsDocument[]> => {
+  const { targetProjectId, year } = params;
+  const startDate = new Date(year, 0, 1);
+  const endDate = new Date(year, 12, 0);
+  const visitsByMonth = await Visits.aggregate([
+    {
+      $match: {
+        $and: [
+          { projectId: Types.ObjectId(targetProjectId) },
+          { date: { $gte: startDate, $lte: endDate } },
+        ],
+      },
+    },
+    {
+      $group: {
+        _id: {
+          projectId: '$projectId',
+          year: { $year: '$date' },
+          month: { $month: '$date' },
+          ip: '$ip',
+        },
+      },
+    },
+    {
+      $group: {
+        _id: {
+          projectId: '$_id.projectId',
+          year: '$_id.year',
+          month: '$_id.month',
+        },
+        count: { $sum: 1 },
+      },
+    },
+    { $sort: { '_id.month': 1 } },
+  ]);
+  return visitsByMonth;
+};
+
+interface FillMonthCountParams {
+  visitsByMonth: IVisitsDocument[];
+  targetProjectId: string;
+  year: number;
+}
+
+const fillMonthCountsWithZero = (params: FillMonthCountParams) => {
+  const { visitsByMonth, targetProjectId: projectId, year }: FillMonthCountParams = params;
   const lastMonth = 12;
   let i = 0;
   const filledVisitsByMonth = Array(lastMonth)
@@ -112,7 +129,7 @@ const fillMonthCountsWithZero = (visitsByMonth: IVisitsDocument[], year: number)
         return targetData;
       }
       const zeroData = {
-        _id: { year, month: targetMonth },
+        _id: { year, month: targetMonth, projectId },
         count: 0,
       };
       return zeroData;
